Group My Account controllers under namespace
To follow the approach taken through the application on grouping
controllers related with a domain under a namespace for the domain.

In this case multiple "my_account_*" controllers are turned into
"my_account/*".
