+++ cosmetics-web/app/controllers/my_account/name_controller.rb
@@ -1,0 +1,20 @@
+module MyAccount
+  class NameController < ApplicationController
+    def show
+      @user = current_user
+      render "name"
+    end
+
+    def update
+      @user = current_user
+
+      @user.name = dig_params(:name)
+
+      if @user.save
+        redirect_to my_account_path, confirmation: "Name changed successfully"
+      else
+        render "name"
+      end
+    end
+  end
+end
