+++ cosmetics-web/app/views/my_account/email/email.html.erb
