+++ cosmetics-web/app/views/my_account/password/password.html.erb
